Move Hash class to separate file
